Split unit tests out into separate files
